feat(backup): set a backup target default
The first backup target will become the default backup target
whether the `Default` in `Spec` is set `true` or not. If it is not
the first backup target with the `Default` in `Spec` set `true`,
backup target controller will set `Default` false both in `Spec`
and `Status` for previous default backup target and then setup new
backup target as a default one.

Ref: 5411

Signed-off-by: James Lu <james.lu@suse.com>